+++ libs/core/src/lib/select/option/option.component.ts
@@ -3,14 +3,19 @@ import {
     ChangeDetectorRef,
     Component,
     ElementRef,
-    EventEmitter,
     HostBinding,
     HostListener,
     Input,
+    OnDestroy,
     OnInit,
-    Output,
     ViewEncapsulation
 } from '@angular/core';
+import { filter, map } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
+import { KeyUtil } from '../../utils/functions/key-util';
+import { SelectProxy } from '../select-proxy.service';
+
+let optionUniqueId: number = 0;
 
 /**
  * Used to represent an option of the select component.
@@ -18,24 +23,35 @@ import {
 @Component({
     selector: 'fd-option',
     templateUrl: './option.component.html',
-    styleUrls: ['./option.component.scss'],
     encapsulation: ViewEncapsulation.None,
+    changeDetection: ChangeDetectionStrategy.OnPush,
     host: {
-        '[class.fd-option-default-custom]': 'true',
+        'class': 'fd-list__item',
         '[attr.aria-disabled]': 'disabled',
+        '[attr.aria-selected]': 'selected',
+        '[attr.id]': 'id',
         '[tabindex]': 'disabled ? -1 : 0',
         role: 'option'
     },
-    changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class OptionComponent implements OnInit {
-    /** @hidden */
-    @HostBinding('class.fd-menu__item')
-    fdMenuItemClass: boolean = true;
+    styles: [`
+        .fd-list__item[aria-disabled="true"],
+        .fd-list__item.is-disabled,
+        .fd-list__item:disabled {
+            opacity: 0.4;
+            cursor: not-allowed;
+        }
 
-    /** @hidden */
-    @HostBinding('class.is-selected')
-    selected: boolean = false;
+        .fd-list__item:disabled:focus {
+            outline: none;
+        }
+    `]
+
+})
+export class OptionComponent implements OnInit, OnDestroy {
+
+    /** Option id attribute */
+    @Input()
+    id: string = `fd-option-${optionUniqueId++}`;
 
     /** Value of the option. Similar to how a native select operates. */
     @Input()
@@ -45,56 +61,81 @@ export class OptionComponent implements OnInit {
     @Input()
     disabled: boolean = false;
 
-    /** Override for the view value of the option. If none is provided, the text content is used. */
-    @Input()
-    viewValue: string;
-
-    /** Emitted when the selected state changes. */
-    @Output()
-    readonly selectedChange: EventEmitter<OptionComponent> = new EventEmitter<OptionComponent>();
+    /** @hidden */
+    @HostBinding('class.is-selected')
+    selected: boolean = false;
 
     /** @hidden */
-    constructor(private _elRef: ElementRef, private _changeDetRef: ChangeDetectorRef) {}
+    private _subscriptions: Subscription = new Subscription();
+
+    /** @hidden */
+    @HostListener('click')
+    @HostListener('keydown', ['$event'])
+    selectionHandler(event?: KeyboardEvent): void {
+        if (!event || event && KeyUtil.isKey(event, [' ', 'Enter'])) {
+            if (event) {
+                event.preventDefault();
+                event.stopPropagation();
+            }
+            if (!this.disabled) {
+                this.setSelected(true, true);
+            }
+        }
+    }
+
+    /** @hidden */
+    constructor(
+        private _elementRef: ElementRef,
+        private _selectProxy: SelectProxy,
+        private _changeDetRef: ChangeDetectorRef
+    ) {}
 
     /** @hidden */
     ngOnInit(): void {
-        if (this.selected && !this.disabled) {
-            this.focus();
-        }
+        this._observeValue();
+    }
+
+    /** @hidden */
+    ngOnDestroy(): void {
+        this._subscriptions.unsubscribe();
     }
 
     /** Returns the view value text of the option, or the viewValue input if it exists. */
     get viewValueText(): string {
-        return this.viewValue ? this.viewValue : ((this._elRef.nativeElement as HTMLElement).textContent || '').trim();
+        return (this._elementRef.nativeElement.textContent || '').trim();
     }
 
-    /** Returns the view value text of the option, or the viewValue input if it exists. */
-    setSelected(value: boolean, fireEvent: boolean = true): void {
+    /** Set control selected state
+     * @param value - whether is selected
+     * @param controlChange - whether is a change initially coming from OptionComponent
+     * */
+    setSelected(value: boolean, controlChange: boolean): void {
         this.selected = value;
 
-        if (fireEvent) {
-            this.selectedChange.emit(this);
+        if (value) {
+            this._selectProxy.optionStateChange$.next({option: this, controlChange: controlChange})
         }
+        this._changeDetRef.markForCheck();
     }
 
     /** Focuses the element. */
     focus(): void {
-        (this._elRef.nativeElement as HTMLElement).focus();
+        (this._elementRef.nativeElement as HTMLElement).focus();
     }
 
     /** Returns HTMLElement representation of the component. */
     getHtmlElement(): HTMLElement {
-        return this._elRef.nativeElement as HTMLElement;
+        return this._elementRef.nativeElement as HTMLElement;
     }
 
     /** @hidden */
-    @HostListener('keydown.enter')
-    @HostListener('click')
-    selectionHandler(): void {
-        if (!this.selected && !this.disabled) {
-            this.selected = true;
-            this._changeDetRef.markForCheck();
-            this.selectedChange.emit(this);
-        }
+    private _observeValue(): void {
+        this._subscriptions.add(
+            this._selectProxy.value$.asObservable()
+                .pipe(
+                    filter(value => value !== undefined),
+                    map(value => value === this.value)
+                ).subscribe(isSelected => this.setSelected(isSelected, false))
+        );
     }
 }